pyproject.toml
[project]
name = "echo-mcp"
version = "0.1.0"
description = "ECHO — Explore Calls, Hearings & Observations. MCP server for searching Zoom meeting transcripts."
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
echo-mcp = "echo_mcp.mcp_server:main"
echo-login = "echo_mcp.cli:login_cli"
echo-logout = "echo_mcp.cli:logout_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
